requirements.txt
crewai
crewai_tools
langchain_google_genai
python-dotenv
serpapi
streamlit
